Renamed existing `windows.py` to `views.py`
